+++ src/services/App.php
@@ -52,6 +52,11 @@ class App extends Component
      */
     public $orderStatuses;
 
+    /**
+     * @var Emails
+     */
+    public $emails;
+
     public function init()
     {
         $this->settings = new Settings();
@@ -62,5 +67,6 @@ class App extends Component
         $this->messages = new Messages();
         $this->customers = new Customers();
         $this->orderStatuses = new OrderStatuses();
+        $this->emails = new Emails();
     }
 }